Alterando formatos de retorno de int para size_t

+++ utils/check_type/ft_isalpha.c
@@ -6,13 +6,13 @@
 /*   By: apaula-b <apaula-b@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 19:19:42 by apaula-b          #+#    #+#             */
-/*   Updated: 2021/06/20 13:23:32 by apaula-b         ###   ########.fr       */
+/*   Updated: 2021/06/20 18:54:58 by apaula-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../ft_printf.h"
 
-int	ft_isalpha(int c)
+size_t	ft_isalpha(size_t c)
 {
 	return ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z'));
 }
